refactor: move AdminLayout to right folder
