+++ src/app/(admin)/AdminLayout/sidebar/data/ui-elements-list.ts
