+++ src/app/(admin)/globals.css
