+++ src/app/(admin)/layout.tsx
@@ -1,15 +1,39 @@
 'use client';
-import React from 'react';
+import { Geist, Geist_Mono } from "next/font/google";
+import "./globals.css";
+import { Provider } from 'react-redux';
+import { store } from '../../application/redux/store';
 
-export default function AdminLayout({
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export default function RootLayout({
   children,
-}: {
+}: Readonly<{
   children: React.ReactNode;
-}) {
+}>) {
   return (
-    <div className="min-h-screen bg-gray-50">
-      {children}
-    </div>
+    <html lang="vi" suppressHydrationWarning>
+      <head>
+        <title>🎬 Studio Booking - Thuê Studio Chuyên Nghiệp</title>
+        <meta name="description" content="Hệ thống đặt lịch thuê studio chuyên nghiệp cho chụp ảnh, quay phim, livestream và workshop." />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+      </head>
+      <body 
+        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+        suppressHydrationWarning
+      >
+        <Provider store={store}>
+          {children}
+        </Provider>
+      </body>
+    </html>
   );
 }
-
