--- src/app/admin/page.tsx
@@ -1,12 +1,0 @@
-export default function AdminPage() {
-  return (
-    <div>
-      <h1 className="text-2xl font-bold">Admin Dashboard</h1>
-
-      <p className="mt-4 text-sm text-muted-foreground">
-        This is a placeholder admin page. The AdminLayout (sidebar + header)
-        should wrap this content.
-      </p>
-    </div>
-  );
-}
